% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean-codes.R
\name{clean_codes}
\alias{clean_codes}
\title{Standardise ICD-10 and ATC codes}
\usage{
clean_codes(x)
}
\arguments{
\item{x}{A character vector or factor containing ICD-10 or ATC codes.}
}
\value{
A character vector containing standardised codes.
}
\description{
Converts ICD-10 or ATC codes to uppercase and removes spaces,
punctuation, and other non-alphanumeric characters.
}
\examples{
clean_codes(
  c(
    "i21.0",
    " I50 ",
    "a10ba02"
  )
)

}
